+++ types/disruptiveTechnologies/temperature/uplink.js
@@ -10,8 +10,14 @@ function consume(event) {
   } else if (eventType === "networkStatus") {
     sample.signalStrength = event.data.networkStatus.signalStrength;
     sample.rssi = event.data.networkStatus.rssi;
-    sample.transmissionMode =
-      event.data.networkStatus.transmissionMode;
+    sample.transmissionMode = event.data.networkStatus.transmissionMode;
+    if (sample.rssi >= -50) {
+      sample.sqi = 3;
+    } else if (sample.rssi < -50 && sample.rssi >= -100) {
+      sample.sqi = 2;
+    } else {
+      sample.sqi = 1;
+    }
     topic = "network_status";
   } else if (eventType === "batteryStatus") {
     sample.batteryLevel = event.data.batteryStatus.percentage;