+++ types/ewattch/squid-hc-lorawan/uplink.spec.js
@@ -5,7 +5,7 @@ const utils = require("test-utils");
 
 const { assert } = chai;
 
-describe("Decentlab PR26 Uplink", () => {
+describe("Ewattch squid hc lorawan", () => {
   let defaultSchema = null;
   let consume = null;
   before((done) => {
@@ -30,10 +30,12 @@ describe("Decentlab PR26 Uplink", () => {
   });
 
   describe("consume()", () => {
-    it("Should decode Decentlab PR26 payload", () => {
+    it("should decode the ewattch squid hc lorawan default uplink", () => {
       const data = {
         data: {
-          payloadHex: "020167000345cb60170c7f",
+          port: 1,
+          payloadHex:
+            "002548509F06A03E0D407D1AF56900EAD300D4A701509F06A03E0D407D1AF56900EAD300D4A701",
         },
       };
 
@@ -43,22 +45,42 @@ describe("Decentlab PR26 Uplink", () => {
         assert.typeOf(value.data, "object");
 
         assert.equal(value.topic, "default");
-        assert.equal(value.data.pressure, 0.045257568359375);
-        assert.equal(value.data.temperature, 25.671875);
-        assert.equal(value.data.level, 0.46148229182599165);
+        assert.equal(value.data.channel1, 4.34);
+        assert.equal(value.data.channel2, 8.68);
+        assert.equal(value.data.channel3, 17.36);
+        assert.equal(value.data.channel4, 0.27125);
+        assert.equal(value.data.channel5, 0.5425);
+        assert.equal(value.data.channel6, 1.085);
+        assert.equal(value.data.channel7, 4.34);
+        assert.equal(value.data.channel8, 8.68);
+        assert.equal(value.data.channel9, 17.36);
+        assert.equal(value.data.channel10, 0.27125);
+        assert.equal(value.data.channel11, 0.5425);
+        assert.equal(value.data.channel12, 1.085);
 
         validate(value.data, defaultSchema, { throwError: true });
       });
+
+      consume(data);
+    });
+
+    it("should decode the ewattch squid hc lorawan status uplink", () => {
+      const data = {
+        data: {
+          port: 1,
+          payloadHex: "100A00080204010408083C00",
+        },
+      };
 
       utils.expectEmits((type, value) => {
         assert.equal(type, "sample");
         assert.isNotNull(value);
         assert.typeOf(value.data, "object");
 
-        assert.equal(value.topic, "lifecycle");
-        assert.equal(value.data.voltage, 3.199);
-        assert.equal(value.data.protocolVersion, 2);
-        assert.equal(value.data.deviceID, 359);
+        assert.equal(value.topic, "status");
+        assert.equal(value.data.batteryStatus, "NORMAL");
+        assert.equal(value.data.firmware, "1.4");
+        assert.equal(value.data.periodicity, 60);
 
         validate(value.data, lifecycleSchema, { throwError: true });
       });